+++ .idea/gradle.xml
@@ -9,7 +9,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/beforeafterslider" />
+            <option value="$PROJECT_DIR$/slidetorevealimage" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
