+++ .idea/modules.xml
@@ -4,7 +4,7 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/before_after_slider.iml" filepath="$PROJECT_DIR$/before_after_slider.iml" />
-      <module fileurl="file://$PROJECT_DIR$/beforeafterslider/beforeafterslider.iml" filepath="$PROJECT_DIR$/beforeafterslider/beforeafterslider.iml" />
+      <module fileurl="file://$PROJECT_DIR$/slidetorevealimage/slidetorevealimage.iml" filepath="$PROJECT_DIR$/slidetorevealimage/slidetorevealimage.iml" />
     </modules>
   </component>
 </project>
