+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':beforeafterslider'
+include ':app', ':slidetorevealimage'
