+++ slidetorevealimage/.gitignore
