+++ slidetorevealimage/build.gradle
