+++ slidetorevealimage/proguard-rules.pro
