+++ slidetorevealimage/src/main/AndroidManifest.xml
