+++ slidetorevealimage/src/main/java/it/marcomirisola/extensions/Extensions.kt
@@ -1,4 +1,4 @@
-package com.github.developer__.extensions
+package it.marcomirisola.extensions
 
 import android.graphics.drawable.Drawable
 import android.view.View
