+++ slidetorevealimage/src/main/res/layout/slider_layout.xml
