+++ slidetorevealimage/src/main/res/mipmap/white_circle.png
