+++ slidetorevealimage/src/main/res/values/attr.xml
