+++ slidetorevealimage/src/main/res/values/strings.xml
